Split public SDK facade into package
