+++ astrid/sdk/dto.py
@@ -1,0 +1,45 @@
+"""Public SDK DTOs, result types, and schema record re-exports.
+
+Consolidates the lightweight data-transfer objects that ``import astrid``
+exposes without loading execution machinery.
+"""
+
+from __future__ import annotations
+
+from astrid.contracts.exec_error import ExecError
+from astrid.contracts.schema import (
+    AliasRecord,
+    CapabilityHandle,
+    Output,
+    Port,
+    Provenance,
+    SafetyDeclaration,
+)
+from astrid.core.task.event_stream import (
+    EventStreamRecord,
+)
+from astrid.sdk_results import (
+    Capability,
+    CapabilityType,
+    DiscoveryResult,
+    InvocationResult,
+    _json_safe,
+    _json_safe_mapping,
+)
+
+__all__ = [
+    "AliasRecord",
+    "Capability",
+    "CapabilityHandle",
+    "CapabilityType",
+    "DiscoveryResult",
+    "EventStreamRecord",
+    "ExecError",
+    "InvocationResult",
+    "Output",
+    "Port",
+    "Provenance",
+    "SafetyDeclaration",
+    "_json_safe",
+    "_json_safe_mapping",
+]
